🔗 Adds href to see and seealso
Visual Studio 2019 (16.8+) supports href natively in the IDE,
rendering the hyperlinks in tooltips that allow easy navigation.

This adds the property as a first-class member so it's easier to
discover, and a test using the delegate visitor shows how it
can be used to validate links in your docs.

Fixes #35