@ Add support for custom attributes in XmlVisitor
When visiting to generate/update an XML doc, we were losing the custom attributes that could have been applied to elements.

By making all typed properties a direct pass-through to the dictionary of custom attributes, we simplify significantly the XmlVisitor, while at the same time ensuring both are always in sync. This means reading and then modifying the in-memory model and writing back, should result in a fairly faithful rendering of the original + changes. (no exact round-tripping including formatting is ensured though, for now, seems unnecessary).

Fully custom elements also are preserved as-is, uncluding its original attributes, and the API of the UnknownElement will now throw if you try to modify its Attributes dictionary, since it's just exposing the underlying XElement, which should be modified itself instead.

Fixes #37.